[feat]: Markdown generation through multiple resource types at once
There is now a resource.json file instead of the provider.json.
This file is working as a tree with each resource type as a key of the root dict and a value of the list of the wanted providers.
There is also a default key in case a resource type is specified without providers.
The script wil generate markdown files for the providers specified for a resource type.
The list.json is generated accordingly