[K9VULN-5659] Write script to generate markdown files (#168)
* Adding the markdown generation draft script from Bahar

The script takes a few arguments but two are required:
	- provider_list: the list of providers we want the rules from (json)
	- input_path is the folder we want the rules from
Under the input_path, the script runs through every provider folder and then through each rule for the provider. For each rule, it copies the content of the positive and negative tests and uses the metadata.json file to create a markdown. If there is no markdown, the rule is ignored.

Refs: K9VULN-5659

* [refactor]: Added a max-examples argument to limit the number of compliant and non-compliant examples

For each rule, we copy at most max-examples of compliant examples and max-examples of non-compliant examples in the markdown generated.
This is passed as an argument when calling the local-gen.py

Refs: K9VULN-5659

* feat: Markdown generation through multiple resource types at once

There is now a resource.json file instead of the provider.json.
This file is working as a tree with each resource type as a key of the root dict and a value of the list of the wanted providers.
There is also a default key in case a resource type is specified without providers.
The script wil generate markdown files for the providers specified for a resource type.
The list.json is generated accordingly

* style: Several style changes

The code snippets are added with the correct resource type
Corrected the imports giving them a line each
used islice instead of manually using ints
optimised markdown build using iglob instead of glob
used walrus operator to shorten the get_code_snippets function
replaced all the single quotes with double ones
compliant code examples will arrive before the non-compliant ones

Refs: K9VULN-5659

* feat: added the frontmatter generation

Another argument to give the path to the frontmatter file has been added
The frontmatter is generated correctly

Refs: K9VULN-5659

* refactor: removed outdated attributes' fetch from metadata

Refs: K9VULN-5659

* feat: added the generated documentation

Generated the documentation using the python script

Refs: K9VULN-5659

* style: Applied PR comments to have the code clearer

Refs: K9VULN-5659

* style: Added README and removed outdated script header

Refs:K9VULN-5659

* style: Added platform and display name

Refs: K9VULN-5659

* style: added a line between each metadata info

This line makes the display of the metadata clearer

Refs: K9VULN-5659