[Modules POC] (#185)
* feat: initial local module source parsing

* feat: auto generate equivalent mapping

This mapping will be used for basic module support using
`get_module_equivalent_key`

* fix: Split the Source attribute of ParsedModules in two

Added AbsSource which is used by the go files to create the mapping.
AbsSource is absoulute path to the module from repo root.
The Source attribute will still be used by the Rego.
It is the one directly parsed from the hcl.

* Module poc test suite (#212)

* feat: Added a few terraform modules that will be used for the tests

also moved modules to module folder and created an instance for each

and updated versioning on test modules

Refs: K9VULN-6981

* feat: added new tests for ideal module support

Test for conditional expressions
Test for templating
Test for remote nested module

Refs: K9VULN-6981

* refactor: changed local nested module test

used the auto_minor_version_upgrade which is simpler
also renamed the folder

Refs: K9VULN-6981

* feat: added the negative tests

Refs: K9VULN-6981

* feat: Added expected results to tests

Refs: K9VULN-6981

* feat: added a more complex conditional test

Refs: K9VULN-6981

* style: cleaned up some code

Refs: K9VULN-6983

* fix: reset kics.go file

* Added a few logs to the modules POC (#216)

* feat: added a few logs to the modules POC

Refs: K9VULN-6987

* refactor: Changed the way we check for module before logging

Refs: K9VULN-6987

* Module POC Cleanup (#214)

* fixup: removed also from modules_test.go

Refs: K9VULN-6983

* fix: added missing argument in scan function

Refs: K9VULN-6983

* refactor: moved modulesto terraform parse folder

also removed useless function
and changed PascalCase to camelCase for private methods

Refs: K9VULN-6983

---------

Co-authored-by: Akim Sadaoui <akim.sadaoui@datadoghq.com>
Co-authored-by: Benjamin Chouraqui <benjamin.chouraqui@datadoghq.com>