fix: use worker pool for analyzer
The analyzer was starting an unlimited amount of goroutine to handle each
individual file, resulting in a huge number of short lived concurrent
goroutine thrashing the scheduler because of the high churn.
With this we instead start a fixed sized worker pool (based on
GOMAXPROCS to correlate it to the actual pod CPU limit).

Refs: K9VULN-8537